REFACT: Separate module for Ad Operator get-set methods
